backend/requirements.txt
fastapi==0.115.6 
loguru==0.7.0
redis==5.0.1
uvicorn==0.21.1
sqlalchemy_utils==0.41.1
pydantic==2.10.4
boto3==1.35.87
python-decouple==3.8
pydantic-settings==2.7.0
kafka==1.3.5